+++ config/importmap.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-pin 'controllers/rpi_turnstile/turnstile_controller'